Morally righteous wretches
Playing the victim from as a soft antag role is tired. It’s fun for one person and draining for everyone else. No one wants to chase down an outlaw just to hear they the one they're hunting is actually righteous and killed Fantasy Jeffrey Epstein.
If you're going to play an antag, own it, or pay the price.